+++ core/src/com/unciv/UncivGame.kt
@@ -466,7 +466,7 @@ open class UncivGame(val isConsoleMode: Boolean = false) : Game(), PlatformSpeci
 
     companion object {
         //region AUTOMATICALLY GENERATED VERSION DATA - DO NOT CHANGE THIS REGION, INCLUDING THIS COMMENT
-        val VERSION = Version("4.10.22", 982)
+        val VERSION = Version("4.11.2", 985)
         //endregion
 
         lateinit var Current: UncivGame