+++ core/src/com/unciv/logic/civilization/Civilization.kt
@@ -45,6 +45,7 @@ import com.unciv.models.ruleset.nation.Personality
 import com.unciv.models.ruleset.tech.Era
 import com.unciv.models.ruleset.tile.ResourceSupplyList
 import com.unciv.models.ruleset.tile.ResourceType
+import com.unciv.models.ruleset.tile.TileImprovement
 import com.unciv.models.ruleset.tile.TileResource
 import com.unciv.models.ruleset.unique.StateForConditionals
 import com.unciv.models.ruleset.unique.TemporaryUnique
@@ -499,7 +500,7 @@ class Civilization : IsPartOfGameInfoSerialization {
     fun getTriggeredUniques(
         trigger: UniqueType,
         stateForConditionals: StateForConditionals = StateForConditionals(this)
-    ) : Sequence<Unique> = sequence {
+    ) : Iterable<Unique> = sequence {
         yieldAll(nation.uniqueMap.getTriggeredUniques(trigger, stateForConditionals))
         yieldAll(cities.asSequence()
             .flatMap { city -> city.cityConstructions.builtBuildingUniqueMap.getTriggeredUniques(trigger, stateForConditionals) }
@@ -512,7 +513,7 @@ class Civilization : IsPartOfGameInfoSerialization {
         yieldAll(tech.techUniques.getTriggeredUniques(trigger, stateForConditionals))
         yieldAll(getEra().uniqueMap.getTriggeredUniques (trigger, stateForConditionals))
         yieldAll(gameInfo.ruleset.globalUniques.uniqueMap.getTriggeredUniques(trigger, stateForConditionals))
-    }
+    }.toList() // Triggers can e.g. add buildings which contain triggers, causing concurrent modification errors
 
     fun matchesFilter(filter: String): Boolean {
         return MultiFilter.multiFilter(filter, ::matchesSingleFilter)
@@ -546,6 +547,22 @@ class Civilization : IsPartOfGameInfoSerialization {
             if (building.replaces == baseBuilding.name)
                 return building
         return baseBuilding
+    }
+
+    fun getEquivalentTileImprovement(tileImprovementName: String): TileImprovement {
+        val tileImprovement = gameInfo.ruleset.tileImprovements[tileImprovementName]
+            ?: throw UncivShowableException("Improvement $tileImprovementName doesn't seem to exist!")
+        return getEquivalentTileImprovement(tileImprovement)
+    }
+
+    fun getEquivalentTileImprovement(tileImprovement: TileImprovement): TileImprovement {
+        if (tileImprovement.replaces != null)
+            return getEquivalentTileImprovement(tileImprovement.replaces!!)
+
+        for (improvement in cache.uniqueImprovements)
+            if (improvement.replaces == tileImprovement.name)
+                return improvement
+        return tileImprovement
     }
 
     fun getEquivalentUnit(baseUnitName: String): BaseUnit {