+++ core/src/com/unciv/logic/civilization/managers/PolicyManager.kt
@@ -117,6 +117,12 @@ class PolicyManager : IsPartOfGameInfoSerialization {
         }
     }
 
+    private fun removePolicyFromTransients(policy: Policy) {
+        for (unique in policy.uniqueObjects) {
+            policyUniques.removeUnique(unique)
+        }
+    }
+
     fun addCulture(culture: Int) {
         val couldAdoptPolicyBefore = canAdoptPolicy()
         storedCulture += culture
@@ -230,6 +236,30 @@ class PolicyManager : IsPartOfGameInfoSerialization {
         }
 
         if (!canAdoptPolicy()) shouldOpenPolicyPicker = false
+    }
+
+    fun removePolicy(policy: Policy, branchCompletion: Boolean = false) {
+        if (!branchCompletion)
+            numberOfAdoptedPolicies--
+
+        adoptedPolicies.remove(policy.name)
+        removePolicyFromTransients(policy)
+
+        // if a branch is already marked as complete, revert it to incomplete
+        if (!branchCompletion) {
+            val branch = policy.branch
+            if (branch.policies.count { isAdopted(it.name) } == branch.policies.size - 1) {
+                removePolicy(branch.policies.last(), true)
+            }
+        }
+
+        civInfo.cache.updateCivResources()
+
+        // This ALSO has the side-effect of updating the CivInfo statForNextTurn so we don't need to call it explicitly
+        for (city in civInfo.cities) {
+            city.cityStats.update()
+            city.reassignPopulationDeferred()
+        }
     }
 
     /**