+++ core/src/com/unciv/models/ruleset/validation/UniqueValidator.kt
@@ -172,12 +172,14 @@ class UniqueValidator(val ruleset: Ruleset) {
                 continue
 
             rulesetErrors.add(
-                "$prefix contains the conditional \"${conditional.text}\"." +
+                "$prefix contains conditional \"${conditional.text}\"." +
                 " This contains the parameter ${complianceError.parameterName} which does not fit parameter type" +
                 " ${complianceError.acceptableParameterTypes.joinToString(" or ") { it.parameterName }} !",
                 complianceError.errorSeverity.getRulesetErrorSeverity(), uniqueContainer, unique
             )
         }
+
+        addDeprecationAnnotationErrors(conditional, "$prefix contains conditional \"${conditional.text}\" which", rulesetErrors, uniqueContainer)
     }
 
     private fun addDeprecationAnnotationErrors(
@@ -207,6 +209,11 @@ class UniqueValidator(val ruleset: Ruleset) {
         if (unique.type == null) return emptyList()
         val errorList = ArrayList<UniqueComplianceError>()
         for ((index, param) in unique.params.withIndex()) {
+            // Trying to catch the error at #11404
+            if (unique.type.parameterTypeMap.size != unique.params.size) {
+                throw Exception("Unique ${unique.text} has ${unique.params.size} parameters, " +
+                        "but its type ${unique.type} only ${unique.type.parameterTypeMap.size} parameters?!")
+            }
             val acceptableParamTypes = unique.type.parameterTypeMap[index]
             val errorTypesForAcceptableParameters =
                 acceptableParamTypes.map { getParamTypeErrorSeverityCached(it, param) }