+++ core/src/com/unciv/models/stats/Stats.kt
@@ -186,9 +186,9 @@ open class Stats(
     }
 
     /** Return a string of just +/- value and Stat symbol*/
-    fun toStringOnlyIcons(): String {
+    fun toStringOnlyIcons(addPlusSign: Boolean = true): String {
         return this.joinToString {
-            (if (it.value > 0) "+" else "") + it.value.toInt() + " " + it.key.character
+            (if (addPlusSign && it.value > 0) "+" else "") + it.value.toInt() + " " + it.key.character
         }
     }